+++ src/client/client.ts
@@ -1,6 +1,7 @@
 import { Elm } from './elm/Main';
 import { setup as clipboardSetup } from './Clipboard';
 import { setup as storageSetup } from './Storage';
+import { loadCookieConsentChoice, setup as analyticsSetup } from './Analytics';
 
 window.loadApp = (isProduction) => {
   const { clientWidth, clientHeight } = document.documentElement;
@@ -12,9 +13,11 @@ window.loadApp = (isProduction) => {
       },
       isProduction,
       version: __VERSION__,
+      showCookieConsentBanner: loadCookieConsentChoice() === undefined,
     },
   });
 
   storageSetup(app);
   clipboardSetup(app);
+  analyticsSetup(app);
 };